pythonsdk/requirements.txt
numpy>=1.19
protobuf==3.14.0
grpcio>=1.31
grpcio-tools>=1.31
netifaces>=0.10
pandas>=0.22
pyntcloud>=0.1.3
matplotlib>=2.1
opencv-python>=3.2
ipycanvas==0.7.0
